Increasing sine modulation precision

+++ main/sin_calculator.h
@@ -3,6 +3,6 @@
 
 void sin_init_timer(void);
 
-void sin_set_freq(const uint8_t freq_value);
+void sin_set_freq(const float freq_value);
 
 #endif
